refactor: 결제 관련 폴더 이름 변경
