+++ src/api/payment/payment.js
@@ -1,14 +1,15 @@
-import axios from '../axios.js';
+import axios from '@/api/axios'
 
-// 토스페이먼츠 결제 승인
+// 결제 승인 API
 export const confirmPayment = async (paymentInfo) => {
   try {
-    const response = await axios.post('/payment/confirm', paymentInfo);
-    return response.data;
+    const response = await axios.post('/payment/confirm', paymentInfo)
+    return response.data
   } catch (error) {
-    throw error;
+    console.error('결제 승인 API 호출 실패:', error)
+    throw error
   }
-};
+}
 
 // 결제 상태 확인
 export const checkPaymentStatus = async (orderId) => {
